ErrorReport could be nil
The following error occurs in this controller.

A NoMethodError occurred in js_exception_notifier#javascript_error:

  undefined method `[]' for nil:NilClass

-------------------------------
Request:
-------------------------------

  * URL        : https://www.example.com/js_exception_notifier
  * HTTP Method: POST

This error could occur if a post request is send without an param['errorReport']

How could just a request be made is not yet clear to me, but they are happening like 10-20 a day with real users and not with search bots.
As a safety we should check if there is param['errorReport'] send to this url and if not, report for this